+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Release Notes
 ## Unreleased
 
+## Platform Version 0.7.4 - 2020-04-22
+* Added Partitioner trait for assigning partitions based on record keys
+* Deprecated the `TopicProducer::send_record` method
+
 ## Platform Version 0.7.3 - 2020-04-02
 * Added batching for producing records with `send_all` API ([#896](https://github.com/infinyon/fluvio/pull/896)).
 * WASM based Smart Stream Filter MVP ([#901](https://github.com/infinyon/fluvio/pull/901)).